Move wallet to features folder
