+++ idea/frontend/.eslintrc
@@ -42,7 +42,9 @@
 
       "rules": {
         "@typescript-eslint/no-unsafe-enum-comparison": "off",
-        "@typescript-eslint/no-floating-promises": "off"
+        "@typescript-eslint/no-floating-promises": "off",
+        "react/display-name": "off",
+        "import/no-named-as-default": "off"
       }
     }
   ]
