+++ idea/frontend/src/app/providers/modal/consts.ts
@@ -1,5 +1,4 @@
 import { NetworkModal } from '@/widgets/networkModal';
-import { AccountsModal } from '@/widgets/accountsModal';
 import { TransactionModal } from '@/widgets/transactionModal';
 import { UploadFileModal } from '@/widgets/uploadFileModal';
 import { UploadMetadataModal } from '@/widgets/uploadMetadataModal';
@@ -8,7 +7,6 @@ import { TransferBalanceModal } from '@/widgets/transferBalanceModal';
 
 const MODALS = {
   network: NetworkModal,
-  accounts: AccountsModal,
   metadata: UploadMetadataModal,
   uploadFile: UploadFileModal,
   transaction: TransactionModal,
