+++ idea/frontend/src/features/wallet/consts.ts
