+++ idea/frontend/src/features/wallet/index.ts
@@ -1,0 +1,3 @@
+import { Wallet } from './ui';
+
+export { Wallet };
