+++ idea/frontend/src/features/wallet/types.ts
