+++ idea/frontend/src/features/wallet/ui/account-button/account-button.module.scss
