+++ idea/frontend/src/features/wallet/ui/account-button/index.ts
@@ -1,0 +1,3 @@
+import { AccountButton } from './account-button';
+
+export { AccountButton };
