+++ idea/frontend/src/features/wallet/ui/accounts-modal/accountItem/AccountItem.module.scss
