+++ idea/frontend/src/features/wallet/ui/accounts-modal/accountItem/index.ts
