+++ idea/frontend/src/features/wallet/ui/accounts-modal/accountList/AccountList.tsx
