+++ idea/frontend/src/features/wallet/ui/accounts-modal/accounts-modal.module.scss
