+++ idea/frontend/src/features/wallet/ui/accounts-modal/index.ts
@@ -1,0 +1,3 @@
+import { AccountsModal } from './accounts-modal';
+
+export { AccountsModal };
