+++ idea/frontend/src/features/wallet/ui/accounts-modal/wallet/Wallet.module.scss
