+++ idea/frontend/src/features/wallet/ui/accounts-modal/wallet/index.ts
