+++ idea/frontend/src/features/wallet/ui/accounts-modal/wallets/Wallets.module.scss
