+++ idea/frontend/src/features/wallet/ui/accounts-modal/wallets/index.ts
