+++ idea/frontend/src/features/wallet/ui/balance/balance.module.scss
