+++ idea/frontend/src/features/wallet/ui/balance/index.ts
