+++ idea/frontend/src/features/wallet/ui/index.ts
@@ -1,0 +1,3 @@
+import { Wallet } from './wallet';
+
+export { Wallet };
