+++ idea/frontend/src/features/wallet/ui/wallet/wallet.module.scss
@@ -8,11 +8,6 @@
   align-items: center;
 }
 
-.onboardingTooltip {
-  position: relative;
-  margin-left: $margin;
-}
-
 .accountBtn {
   background-color: $bgColorSecondary;
