--- idea/frontend/src/shared/ui/accountButton/index.ts
@@ -1,3 +1,0 @@
-import { AccountButton } from './AccountButton';
-
-export { AccountButton };
