--- idea/frontend/src/widgets/accountsModal/index.ts
@@ -1,3 +1,0 @@
-import { AccountsModal } from './ui/AccountsModal';
-
-export { AccountsModal };
