--- idea/frontend/src/widgets/accountsModal/model/index.ts
@@ -1,5 +1,0 @@
-import { WALLETS, WALLET } from './consts';
-import { WalletId } from './types';
-
-export { WALLETS, WALLET };
-export type { WalletId };
