--- idea/frontend/src/widgets/header/ui/wallet/index.ts
@@ -1,3 +1,0 @@
-import { Wallet } from './Wallet';
-
-export { Wallet };
